// src/lib/generated/types.gen.ts
export type Account = {
    readonly id: number;
    /**
     * Designates that this user has all permissions without explicitly assigning them.
     */
    readonly is_superuser: boolean;
    readonly email: string;
    readonly is_pro_user: boolean;
    readonly is_staff: boolean;
    readonly is_active: boolean;
    readonly is_manager: boolean;
    readonly date_joined: string;
    readonly last_login: string;
    readonly staff_profile: (StaffProfile);
    readonly groups: Array<Group>;
    readonly managers: Array<Manager>;
};
// ...
export type Group = {
    readonly id: number;
    name: string;
    permissions: Array<(number)>;
};

export type Image = {
    readonly id: number;
    image: string;
    title: string;
    title_en?: (string) | null;
    title_nl?: (string) | null;
    title_fr?: (string) | null;
    image_type?: ImageTypeEnum;
    readonly created_at: string;
    readonly updated_at: string;
    owner?: (number) | null;
};
// ...
export type Manager = {
    readonly id: number;
    email: string;
    is_staff?: boolean;
    is_active?: boolean;
    /**
     * Designates that this user has all permissions without explicitly assigning them.
     */
    is_superuser?: boolean;
    is_pro_user?: boolean;
    staff_profile?: (number) | null;
};
// ...
export type StaffProfile = {
    readonly id: number;
    readonly profile_picture: (Image);
    profile_picture_id?: number;
    first_name: string;
    last_name: string;
    preferred_name?: string;
    preferred_weekly_working_hours?: number;
    phone_number?: string;
    address?: string;
    date_of_birth?: (string) | null;
    emergency_contact_name?: string;
    emergency_contact_number?: string;
    emergency_contact_relationship?: string;
    readonly created_at: string;
    readonly updated_at: string;
};